[Chore] #150 placeId type 변경

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/FormPlace.kt
@@ -1,7 +1,7 @@
 package kr.tekit.lion.daongil.domain.model
 
 data class FormPlace (
-    val placeId : Int,
+    val placeId : Long,
     val placeImage : String, // 사진 url
     val placeAddress : String, // 주소
     val placeName : String, // 장소명
